+++ ruby/spec/spec_helper.rb
@@ -1,3 +1,4 @@
 require_relative '../lib/gitaly_server.rb'
 require_relative '../lib/gitlab/git.rb'
+require_relative 'support/sentry.rb'
 require 'test_repo_helper'